Keep disk cache helpers readiness-clean
