+++ abicheck/dumper_cache.py
@@ -1,0 +1,42 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Best-effort AST cache path helpers."""
+
+from __future__ import annotations
+
+import logging
+import os
+import sys
+import tempfile
+from pathlib import Path
+
+log = logging.getLogger(__name__)
+
+
+def _cache_path(key: str, backend: str = "castxml") -> Path:
+    # One sub-directory + file extension per backend so castxml XML and clang
+    # JSON caches live side by side without clashing.
+    ext = "json" if backend == "clang" else "xml"
+    if sys.platform == "win32":
+        local = os.environ.get("LOCALAPPDATA")
+        cache_dir = (
+            Path(local) / "abi_check" / backend
+            if local
+            else Path.home() / "AppData" / "Local" / "abi_check" / backend
+        )
+    else:
+        xdg_cache = os.environ.get("XDG_CACHE_HOME")
+        base = Path(xdg_cache) if xdg_cache else Path.home() / ".cache"
+        cache_dir = base / "abi_check" / backend
+    try:
+        cache_dir.mkdir(parents=True, exist_ok=True)
+    except OSError as exc:
+        fallback = Path(tempfile.gettempdir()) / "abi_check" / backend
+        log.warning(
+            "AST cache directory %s is unavailable (%s); using %s",
+            cache_dir,
+            exc,
+            fallback,
+        )
+        fallback.mkdir(parents=True, exist_ok=True)
+        cache_dir = fallback
+    return cache_dir / f"{key}.{ext}"
