+++ tests/test_source_replay.py
@@ -545,6 +545,7 @@ def test_cache_put_ignores_unwritable_cache_dir() -> None:
     cache = SourceAbiCache(Path("/proc/abicheck-l4-cache"))
 
     cache.put("k", SourceAbiTu(tu_id="cu://x"))
+    assert cache.get("k") is None
 
 
 def test_cache_key_changes_with_argv_forced_include(tmp_path: Path) -> None:
